Add [Flags] attribute to multiple enums in SDL3
This commit introduces the `[Flags]` attribute to several enumerations in the `Hexa.NET.SDL3` namespace. The affected enums include `SDLGPUColorComponentFlags`, `SDLInitFlags`, `SDLKeymod`, `SDLMessageBoxFlags`, `SDLMouseButtonFlags`, and `SDLWindowFlags`. This enhancement allows for bitwise operations on the enum values, enabling the combination of multiple flags in a single variable, thus improving flexibility and usability in scenarios requiring multiple options.